+++ examples/scene-viewer/src/main.rs
@@ -1,6 +1,8 @@
-use glam::{DVec2, UVec2, Vec3, Vec3A};
+use glam::{DVec2, Mat3A, Mat4, UVec2, Vec3, Vec3A};
+use parking_lot::Mutex;
 use pico_args::Arguments;
 use rend3::{
+    format_sso,
     types::{Backend, Camera, CameraProjection, DirectionalLight, Texture, TextureFormat},
     util::typedefs::FastHashMap,
     Renderer,
@@ -23,47 +25,68 @@ use winit::{
 
 mod platform;
 
-fn load_skybox(renderer: &Renderer, routine: &mut PbrRenderRoutine) -> Result<(), Box<dyn std::error::Error>> {
+fn load_skybox(renderer: &Renderer, routine: &Mutex<PbrRenderRoutine>) -> Result<(), Box<dyn std::error::Error>> {
+    profiling::scope!("load skybox");
+
     let name = concat!(env!("CARGO_MANIFEST_DIR"), "/data/skybox.basis");
     let file = std::fs::read(name).unwrap_or_else(|_| panic!("Could not read skybox {}", name));
 
-    let transcoder = basis::Transcoder::new();
-    let image_info = transcoder.get_image_info(&file, 0).ok_or("skybox image missing")?;
+    let mut transcoder = basis_universal::Transcoder::new();
+    let image_info = transcoder.image_info(&file, 0).ok_or("skybox image missing")?;
 
-    let mips = transcoder.get_total_image_levels(&file, 0);
+    let mips = transcoder.image_level_count(&file, 0);
 
-    let mut prepared = transcoder
-        .prepare_transcoding(&file)
-        .ok_or("could not prepare skybox transcoding")?;
-    let mut image = Vec::with_capacity(image_info.total_blocks as usize * 16 * 6);
+    {
+        profiling::scope!("prepare basis transcoding");
+        transcoder
+            .prepare_transcoding(&file)
+            .map_err(|_| "could not prepare skybox transcoding")?
+    }
+    let mut image = Vec::with_capacity(image_info.m_total_blocks as usize * 16 * 6);
     for i in 0..6 {
         for mip in 0..mips {
-            let mip_info = transcoder.get_image_level_info(&file, 0, mip).unwrap();
-            let data = prepared.transcode_image_level(i, mip, basis::TargetTextureFormat::Rgba32)?;
-            image.extend_from_slice(&data[0..(mip_info.orig_width * mip_info.orig_height * 4) as usize]);
+            profiling::scope!("basis transcoding", &format_sso!("layer {} mip {}", i, mip));
+            let mip_info = transcoder.image_level_info(&file, 0, mip).unwrap();
+            let data = transcoder
+                .transcode_image_level(
+                    &file,
+                    basis_universal::TranscoderTextureFormat::RGBA32,
+                    basis_universal::TranscodeParameters {
+                        image_index: i,
+                        level_index: mip,
+                        decode_flags: None,
+                        output_row_pitch_in_blocks_or_pixels: None,
+                        output_rows_in_pixels: None,
+                    },
+                )
+                .map_err(|_| "failed to transcode")?;
+            image.extend_from_slice(&data[0..(mip_info.m_orig_width * mip_info.m_orig_height * 4) as usize]);
         }
     }
-    drop(prepared);
 
     let handle = renderer.add_texture_cube(Texture {
         format: TextureFormat::Rgba8UnormSrgb,
-        size: UVec2::new(image_info.width, image_info.height),
+        size: UVec2::new(image_info.m_width, image_info.m_height),
         data: image,
         label: Some("background".into()),
         mip_count: rend3::types::MipmapCount::Specific(NonZeroU32::new(mips).unwrap()),
         mip_source: rend3::types::MipmapSource::Uploaded,
     });
-    routine.set_background_texture(Some(handle));
+    routine.lock().set_background_texture(Some(handle));
     Ok(())
 }
 
 fn load_gltf(renderer: &Renderer, location: String) -> rend3_gltf::LoadedGltfScene {
+    profiling::scope!("loading gltf");
     let path = Path::new(&location);
     let parent = path.parent().unwrap();
 
-    log::info!("Reading gltf file: {}", path.display());
-    let gltf_data =
-        std::fs::read(&path).unwrap_or_else(|e| panic!("tried to load gltf file {}: {}", path.display(), e));
+    let path_str = path.as_os_str().to_string_lossy();
+    log::info!("Reading gltf file: {}", path_str);
+    let gltf_data = {
+        profiling::scope!("reading gltf file", &path_str);
+        std::fs::read(&path).unwrap_or_else(|e| panic!("tried to load gltf file {}: {}", path_str, e))
+    };
 
     pollster::block_on(rend3_gltf::load_gltf(renderer, &gltf_data, |uri| {
         rend3_gltf::filesystem_io_func(&parent, uri)
@@ -107,12 +130,14 @@ fn main() {
     let desired_mode = args.value_from_fn(["-m", "--mode"], extract_mode).ok();
     let file_to_load: Option<String> = args.free_from_str().ok();
 
-    let event_loop = EventLoop::new();
-
-    let window = {
+    let (event_loop, window) = {
+        profiling::scope!("creating window");
+        let event_loop = EventLoop::new();
         let mut builder = WindowBuilder::new();
         builder = builder.with_title("scene-viewer");
-        builder.build(&event_loop).expect("Could not build window")
+        let window = builder.build(&event_loop).expect("Could not build window");
+
+        (event_loop, window)
     };
 
     let window_size = window.inner_size();
@@ -121,9 +146,15 @@ fn main() {
     let iad = pollster::block_on(rend3::create_iad(desired_backend, desired_device_name, desired_mode)).unwrap();
 
     // The one line of unsafe needed. We just need to guarentee that the window outlives the use of the surface.
-    let surface = unsafe { iad.instance.create_surface(&window) };
+    let surface = unsafe {
+        profiling::scope!("creating surface");
+        iad.instance.create_surface(&window)
+    };
     // Get the preferred format for the surface.
-    let format = surface.get_preferred_format(&iad.adapter).unwrap();
+    let format = {
+        profiling::scope!("getting preferred format");
+        surface.get_preferred_format(&iad.adapter).unwrap()
+    };
     // Configure the surface to be ready for rendering.
     rend3::configure_surface(
         &surface,
@@ -137,23 +168,25 @@ fn main() {
     let renderer = rend3::Renderer::new(iad, Some(window_size.width as f32 / window_size.height as f32)).unwrap();
 
     // Create the pbr pipeline with the same internal resolution and 4x multisampling
-    let mut routine = rend3_pbr::PbrRenderRoutine::new(
+    let routine = Arc::new(Mutex::new(rend3_pbr::PbrRenderRoutine::new(
         &renderer,
         rend3_pbr::RenderTextureOptions {
             resolution: UVec2::new(window_size.width, window_size.height),
-            samples: rend3_pbr::SampleCount::One,
+            samples: rend3_pbr::SampleCount::Four,
         },
         format,
-    );
+    )));
 
+    let routine_clone = Arc::clone(&routine);
     let renderer_clone = Arc::clone(&renderer);
     let _loaded_gltf = std::thread::spawn(move || {
+        profiling::register_thread!("asset loading");
+        load_skybox(&renderer_clone, &routine_clone).unwrap();
         load_gltf(
             &renderer_clone,
             file_to_load.unwrap_or_else(|| concat!(env!("CARGO_MANIFEST_DIR"), "/data/scene.gltf").to_owned()),
         )
     });
-    load_skybox(&renderer, &mut routine).unwrap();
 
     let _directional_light = renderer.add_directional_light(DirectionalLight {
         color: Vec3::ONE,
@@ -163,15 +196,9 @@ fn main() {
     });
     let mut scancode_status = FastHashMap::default();
 
-    let mut camera_location = Camera {
-        projection: CameraProjection::Projection {
-            vfov: 60.0,
-            near: 0.1,
-            pitch: std::f32::consts::FRAC_PI_4,
-            yaw: -std::f32::consts::FRAC_PI_4,
-        },
-        location: Vec3A::new(20.0, 20.0, -20.0),
-    };
+    let mut camera_pitch = std::f32::consts::FRAC_PI_4;
+    let mut camera_yaw = -std::f32::consts::FRAC_PI_4;
+    let mut camera_location = Vec3A::new(20.0, 20.0, -20.0);
 
     let mut previous_profiling_stats: Option<Vec<GpuTimerScopeResult>> = None;
 
@@ -184,6 +211,7 @@ fn main() {
 
     event_loop.run(move |event, _window_target, control| match event {
         Event::MainEventsCleared => {
+            profiling::scope!("MainEventsCleared");
             let now = Instant::now();
 
             let delta_time = now - timestamp_last_frame;
@@ -209,37 +237,32 @@ fn main() {
 
             timestamp_last_frame = now;
 
-            let forward = {
-                if let CameraProjection::Projection { yaw, pitch, .. } = camera_location.projection {
-                    Vec3A::new(yaw.sin() * pitch.cos(), -pitch.sin(), yaw.cos() * pitch.cos())
-                } else {
-                    unreachable!()
-                }
-            };
-            let up = Vec3A::Y;
-            let side: Vec3A = forward.cross(up).normalize();
+            let rotation = Mat3A::from_euler(glam::EulerRot::XYZ, -camera_pitch, -camera_yaw, 0.0).transpose();
+            let forward = rotation.z_axis;
+            let up = rotation.y_axis;
+            let side = -rotation.x_axis;
             let velocity = if button_pressed(&scancode_status, platform::Scancodes::SHIFT) {
                 100.0
             } else {
                 1.0
             };
             if button_pressed(&scancode_status, platform::Scancodes::W) {
-                camera_location.location += forward * velocity * delta_time.as_secs_f32();
+                camera_location += forward * velocity * delta_time.as_secs_f32();
             }
             if button_pressed(&scancode_status, platform::Scancodes::S) {
-                camera_location.location -= forward * velocity * delta_time.as_secs_f32();
+                camera_location -= forward * velocity * delta_time.as_secs_f32();
             }
             if button_pressed(&scancode_status, platform::Scancodes::A) {
-                camera_location.location += side * velocity * delta_time.as_secs_f32();
+                camera_location += side * velocity * delta_time.as_secs_f32();
             }
             if button_pressed(&scancode_status, platform::Scancodes::D) {
-                camera_location.location -= side * velocity * delta_time.as_secs_f32();
+                camera_location -= side * velocity * delta_time.as_secs_f32();
             }
             if button_pressed(&scancode_status, platform::Scancodes::Q) {
-                camera_location.location += up * velocity * delta_time.as_secs_f32();
+                camera_location += up * velocity * delta_time.as_secs_f32();
             }
             if button_pressed(&scancode_status, platform::Scancodes::Z) {
-                camera_location.location -= up * velocity * delta_time.as_secs_f32();
+                camera_location -= up * velocity * delta_time.as_secs_f32();
             }
 
             if button_pressed(&scancode_status, platform::Scancodes::P) {
@@ -292,20 +315,16 @@ fn main() {
             };
 
 
-            if let CameraProjection::Projection { ref mut yaw, ref mut pitch, .. } = camera_location.projection {
-                *yaw += (mouse_delta.x / 1000.0) as f32;
-                *pitch += (mouse_delta.y / 1000.0) as f32;
-                if *yaw < 0.0 {
-                    *yaw += TAU;
-                } else if *yaw >= TAU {
-                    *yaw -= TAU;
-                }
-                *pitch = pitch
-                    .max(-std::f32::consts::FRAC_PI_2 + 0.0001)
-                    .min(std::f32::consts::FRAC_PI_2 - 0.0001);
-            } else {
-                unreachable!()
+            camera_yaw += (mouse_delta.x / 1000.0) as f32;
+            camera_pitch += (mouse_delta.y / 1000.0) as f32;
+            if camera_yaw < 0.0 {
+                camera_yaw += TAU;
+            } else if camera_yaw >= TAU {
+                camera_yaw -= TAU;
             }
+            camera_pitch = camera_pitch
+                .max(-std::f32::consts::FRAC_PI_2 + 0.0001)
+                .min(std::f32::consts::FRAC_PI_2 - 0.0001);
         }
         Event::WindowEvent {
             event: WindowEvent::CloseRequested,
@@ -318,6 +337,7 @@ fn main() {
             event: winit::event::WindowEvent::Resized(size),
             ..
         } => {
+            profiling::scope!("Resized", &format_sso!("{}x{}", size.width, size.height));
             let size = UVec2::new(size.width, size.height);
             // Reconfigure the surface for the new size.
             rend3::configure_surface(
@@ -330,22 +350,36 @@ fn main() {
             // Tell the renderer about the new aspect ratio.
             renderer.set_aspect_ratio(size.x as f32 / size.y as f32);
             // Resize the internal buffers to the same size as the screen.
-            routine.resize(
+            routine.lock().resize(
                 &renderer,
                 rend3_pbr::RenderTextureOptions {
                     resolution: size,
-                    samples: rend3_pbr::SampleCount::One,
+                    samples: rend3_pbr::SampleCount::Four,
                 },
             );
         }
         // Render!
         winit::event::Event::RedrawRequested(..) => {
+        profiling::scope!("RedrawRequested");
             // Update camera
-            renderer.set_camera_data(camera_location);
+
+            let view  = Mat4::from_euler(glam::EulerRot::XYZ, -camera_pitch, -camera_yaw, 0.0);
+            let view = view * Mat4::from_translation((-camera_location).into());
+
+            renderer.set_camera_data(Camera {
+                projection: CameraProjection::Projection {
+                    vfov: 60.0,
+                    near: 0.1,
+                },
+                view
+            });
+
             // Get a frame
             let frame = rend3::util::output::OutputFrame::from_surface(&surface).unwrap();
             // Dispatch a render!
-            previous_profiling_stats = renderer.render(&mut routine, (), frame.as_view());
+            previous_profiling_stats = renderer.render(&mut *routine.lock(), (), frame.as_view());
+            // Present the frame on screen
+            frame.present();
             // mark the end of the frame for tracy/other profilers
             profiling::finish_frame!();
         }