+++ rend3-gltf/src/lib.rs
@@ -17,13 +17,14 @@
 
 use glam::{Mat3, Mat4, UVec2, Vec2, Vec3, Vec4, Vec4Swizzles};
 use gltf::buffer::Source;
+use image::GenericImageView;
 use rend3::{
     types::{self, ObjectHandle},
     util::typedefs::{FastHashMap, SsoString},
     Renderer,
 };
 use rend3_pbr::material;
-use std::{borrow::Cow, collections::hash_map::Entry, future::Future, path::Path};
+use std::{borrow::Cow, collections::hash_map::Entry, future::Future, num::NonZeroU32, path::Path};
 use thiserror::Error;
 
 /// Wrapper around a T that stores an optional label.
@@ -87,8 +88,15 @@ pub struct ImageKey {
     pub srgb: bool,
 }
 
+/// Format
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Texture {
+    pub handle: types::TextureHandle,
+    pub format: types::TextureFormat,
+}
+
 /// Hashmap which stores a mapping from [`ImageKey`] to a labeled handle.
-pub type ImageMap = FastHashMap<ImageKey, Labeled<types::TextureHandle>>;
+pub type ImageMap = FastHashMap<ImageKey, Labeled<Texture>>;
 
 /// A fully loaded Gltf scene.
 #[derive(Debug)]
@@ -110,7 +118,9 @@ pub enum GltfLoadError<E: std::error::Error + 'static> {
     #[error("Texture {0} failed to be loaded from the fs")]
     TextureIo(SsoString, #[source] E),
     #[error("Texture {0} failed to be loaded as an image")]
-    TextureLoad(SsoString, #[source] image::ImageError),
+    TextureDecode(SsoString, #[source] image::ImageError),
+    #[error("Texture {0} failed to be loaded as a ktx2 format due to incompatible format {1:?}")]
+    TextureBadKxt2Format(SsoString, ktx2::Format),
     #[error("Gltf file must have at least one scene")]
     MissingScene,
     #[error("Mesh {0} does not have positions")]
@@ -133,12 +143,15 @@ pub async fn filesystem_io_func(parent_director: impl AsRef<Path>, uri: SsoStrin
         let (_mime, rest) = base64_data.split_once(";").unwrap();
         let (encoding, data) = rest.split_once(",").unwrap();
         assert_eq!(encoding, "base64");
+        profiling::scope!("decoding base64 uri");
         log::info!("loading {} bytes of base64 data", data.len());
         // TODO: errors
         Ok(base64::decode(data).unwrap())
     } else {
         let path_resolved = parent_director.as_ref().join(&*uri);
-        log::info!("loading file '{}' from disk", path_resolved.display());
+        let display = path_resolved.as_os_str().to_string_lossy();
+        profiling::scope!("loading file", &display);
+        log::info!("loading file '{}' from disk", &display);
         std::fs::read(path_resolved)
     }
 }
@@ -167,7 +180,11 @@ where
     Fut: Future<Output = Result<Vec<u8>, E>>,
     E: std::error::Error + 'static,
 {
-    let mut file = gltf::Gltf::from_slice_without_validation(data)?;
+    profiling::scope!("loading gltf");
+    let mut file = {
+        profiling::scope!("parsing gltf");
+        gltf::Gltf::from_slice_without_validation(data)?
+    };
     let blob = file.blob.take();
 
     let buffers = load_buffers(file.buffers(), blob, &mut io_func).await?;
@@ -295,6 +312,7 @@ where
     Fut: Future<Output = Result<Vec<u8>, E>>,
     E: std::error::Error + 'static,
 {
+    profiling::scope!("loading buffers");
     let mut buffers = Vec::with_capacity(file.len());
     let mut blob_index = None;
     for b in file {
@@ -388,6 +406,7 @@ pub fn load_meshes<'a, E: std::error::Error + 'static>(
 
 /// Creates a gltf default material.
 pub fn load_default_material(renderer: &Renderer) -> types::MaterialHandle {
+    profiling::scope!("creating default material");
     renderer.add_material(material::PbrMaterial {
         albedo: material::AlbedoComponent::Value(Vec4::splat(1.0)),
         transparency: material::Transparency::Opaque,
@@ -425,9 +444,13 @@ where
     Fut: Future<Output = Result<Vec<u8>, E>>,
     E: std::error::Error + 'static,
 {
+    profiling::scope!("loading materials and textures");
+
     let mut images = ImageMap::default();
     let mut result = Vec::with_capacity(materials.len());
     for material in materials {
+        profiling::scope!("load material", material.name().unwrap_or_default());
+
         let pbr = material.pbr_metallic_roughness();
         let albedo = pbr.base_color_texture();
         let albedo_factor = pbr.base_color_factor();
@@ -485,11 +508,14 @@ where
         let handle = renderer.add_material(material::PbrMaterial {
             albedo: match albedo_tex {
                 Some(tex) => material::AlbedoComponent::TextureVertexValue {
-                    texture: tex,
+                    texture: tex.handle,
                     value: Vec4::from(albedo_factor),
                     srgb: false,
                 },
-                None => material::AlbedoComponent::Value(Vec4::from(albedo_factor)),
+                None => material::AlbedoComponent::ValueVertex {
+                    value: Vec4::from(albedo_factor),
+                    srgb: false,
+                },
             },
             transparency: match material.alpha_mode() {
                 gltf::material::AlphaMode::Opaque => material::Transparency::Opaque,
@@ -499,21 +525,35 @@ where
                 gltf::material::AlphaMode::Blend => material::Transparency::Blend,
             },
             normal: match normals_tex {
-                Some(tex) => material::NormalTexture::Tricomponent(tex),
-                None => material::NormalTexture::None,
+                Some(tex) if tex.format.describe().components == 2 => material::NormalTexture::Bicomponent(tex.handle),
+                Some(tex) if tex.format.describe().components >= 3 => material::NormalTexture::Tricomponent(tex.handle),
+                _ => material::NormalTexture::None,
             },
             aomr_textures: match (metallic_roughness_tex, occlusion_tex) {
-                (Some(mr), Some(ao)) if mr == ao => material::AoMRTextures::GltfCombined { texture: Some(mr) },
-                (mr, ao) => material::AoMRTextures::GltfSplit {
-                    mr_texture: mr,
-                    ao_texture: ao,
+                (Some(mr), Some(ao)) if mr == ao => material::AoMRTextures::Combined {
+                    texture: Some(mr.handle),
+                },
+                (mr, ao)
+                    if ao
+                        .as_ref()
+                        .map(|ao| ao.format.describe().components < 3)
+                        .unwrap_or(false) =>
+                {
+                    material::AoMRTextures::Split {
+                        mr_texture: util::extract_handle(mr),
+                        ao_texture: util::extract_handle(ao),
+                    }
+                }
+                (mr, ao) => material::AoMRTextures::SwizzledSplit {
+                    mr_texture: util::extract_handle(mr),
+                    ao_texture: util::extract_handle(ao),
                 },
             },
             metallic_factor: Some(metallic_factor),
             roughness_factor: Some(roughness_factor),
             emissive: match emissive_tex {
                 Some(tex) => material::MaterialComponent::TextureValue {
-                    texture: tex,
+                    texture: tex.handle,
                     value: Vec3::from(emissive_factor),
                 },
                 None => material::MaterialComponent::Value(Vec3::from(emissive_factor)),
@@ -545,7 +585,7 @@ pub async fn load_image_cached<F, Fut, E>(
     srgb: bool,
     buffers: &[Vec<u8>],
     io_func: &mut F,
-) -> Result<Labeled<types::TextureHandle>, GltfLoadError<E>>
+) -> Result<Labeled<Texture>, GltfLoadError<E>>
 where
     F: FnMut(SsoString) -> Fut,
     Fut: Future<Output = Result<Vec<u8>, E>>,
@@ -579,13 +619,13 @@ pub async fn load_image<F, Fut, E>(
     srgb: bool,
     buffers: &[Vec<u8>],
     io_func: &mut F,
-) -> Result<Labeled<types::TextureHandle>, GltfLoadError<E>>
+) -> Result<Labeled<Texture>, GltfLoadError<E>>
 where
     F: FnMut(SsoString) -> Fut,
     Fut: Future<Output = Result<Vec<u8>, E>>,
     E: std::error::Error + 'static,
 {
-    // TODO: Address format detection for compressed texs
+    profiling::scope!("load image", image.name().unwrap_or_default());
     let (data, uri) = match image.source() {
         gltf::image::Source::Uri { uri, .. } => {
             let data = io_func(SsoString::from(uri))
@@ -603,28 +643,73 @@ where
         }
     };
 
-    let parsed = image::load_from_memory(&data).map_err(|e| GltfLoadError::TextureLoad(uri, e))?;
-    let rgba = parsed.to_rgba8();
-    let handle = renderer.add_texture_2d(types::Texture {
-        label: image.name().map(str::to_owned),
-        format: match srgb {
-            true => types::TextureFormat::Rgba8UnormSrgb,
-            false => types::TextureFormat::Rgba8Unorm,
-        },
-        size: UVec2::new(rgba.width(), rgba.height()),
-        data: rgba.into_raw(),
-        mip_count: types::MipmapCount::Maximum,
-        mip_source: types::MipmapSource::Generated,
-    });
+    let texture = if let Ok(reader) = ktx2::Reader::new(&data) {
+        profiling::scope!("parsing ktx2");
 
-    Ok(Labeled::new(handle, image.name()))
+        let header = reader.header();
+
+        let src_format = header.format.unwrap();
+        let format = util::map_ktx2_format(src_format, srgb).ok_or_else(|| {
+            GltfLoadError::TextureBadKxt2Format(image.name().map(SsoString::from).unwrap_or_default(), src_format)
+        })?;
+
+        let guaranteed_format = format.describe().guaranteed_format_features;
+        let generate = header.level_count == 1
+            && guaranteed_format.filterable
+            && guaranteed_format.allowed_usages.contains(
+                rend3::types::TextureUsages::TEXTURE_BINDING | rend3::types::TextureUsages::RENDER_ATTACHMENT,
+            );
+
+        types::Texture {
+            label: image.name().map(str::to_owned),
+            format,
+            size: UVec2::new(header.pixel_width, header.pixel_height),
+            data: reader.data().to_vec(),
+            mip_count: if generate {
+                types::MipmapCount::Maximum
+            } else {
+                types::MipmapCount::Specific(NonZeroU32::new(header.level_count).unwrap())
+            },
+            mip_source: if generate {
+                types::MipmapSource::Generated
+            } else {
+                types::MipmapSource::Uploaded
+            },
+        }
+    } else {
+        profiling::scope!("decoding image");
+        let parsed = image::load_from_memory(&data).map_err(|e| GltfLoadError::TextureDecode(uri, e))?;
+        let size = UVec2::new(parsed.width(), parsed.height());
+        let (data, format) = util::convert_dynamic_image(parsed, srgb);
+
+        types::Texture {
+            label: image.name().map(str::to_owned),
+            format,
+            size,
+            data,
+            mip_count: types::MipmapCount::Maximum,
+            mip_source: types::MipmapSource::Generated,
+        }
+    };
+    let format = texture.format;
+    let handle = renderer.add_texture_2d(texture);
+
+    Ok(Labeled::new(Texture { handle, format }, image.name()))
 }
 
 /// Implementation utilities.
 pub mod util {
     use std::future::Future;
 
-    use crate::Labeled;
+    use image::{buffer::ConvertBuffer, Bgra, ImageBuffer, Luma, Rgba};
+    use rend3::types;
+
+    use crate::{Labeled, Texture};
+
+    /// Turns an `Option<Texture>` into `Option<types::TextureHandle>`
+    pub fn extract_handle(texture: Option<Texture>) -> Option<types::TextureHandle> {
+        texture.map(|t| t.handle)
+    }
 
     /// Turns a `Option<Future<Output = Result<Labeled<T>, E>>>>` into a `Future<Output = Result<Option<T>, E>>`
     ///
@@ -641,5 +726,310 @@ pub mod util {
         } else {
             Ok(None)
         }
+    }
+
+    pub fn convert_dynamic_image(image: image::DynamicImage, srgb: bool) -> (Vec<u8>, rend3::types::TextureFormat) {
+        use rend3::types::TextureFormat as r3F;
+
+        profiling::scope!("convert dynamic image");
+        match image {
+            image::DynamicImage::ImageLuma8(i) => (i.into_raw(), r3F::R8Unorm),
+            image::DynamicImage::ImageLumaA8(i) => (
+                ConvertBuffer::<ImageBuffer<Luma<u8>, Vec<u8>>>::convert(&i).into_raw(),
+                r3F::R8Unorm,
+            ),
+            image::DynamicImage::ImageRgb8(i) => (
+                ConvertBuffer::<ImageBuffer<Rgba<u8>, Vec<u8>>>::convert(&i).into_raw(),
+                if srgb { r3F::Rgba8UnormSrgb } else { r3F::Rgba8Unorm },
+            ),
+            image::DynamicImage::ImageRgba8(i) => {
+                (i.into_raw(), if srgb { r3F::Rgba8UnormSrgb } else { r3F::Rgba8Unorm })
+            }
+            image::DynamicImage::ImageBgr8(i) => (
+                ConvertBuffer::<ImageBuffer<Bgra<u8>, Vec<u8>>>::convert(&i).into_raw(),
+                if srgb { r3F::Bgra8UnormSrgb } else { r3F::Bgra8Unorm },
+            ),
+            image::DynamicImage::ImageBgra8(i) => {
+                (i.into_raw(), if srgb { r3F::Bgra8UnormSrgb } else { r3F::Bgra8Unorm })
+            }
+            i => (
+                i.into_rgba8().into_raw(),
+                if srgb { r3F::Rgba8UnormSrgb } else { r3F::Rgba8Unorm },
+            ),
+        }
+    }
+
+    /// Maps a ktx2 format into the rend3's TextureFormat
+    pub fn map_ktx2_format(format: ktx2::Format, srgb: bool) -> Option<rend3::types::TextureFormat> {
+        use ktx2::Format as k2F;
+        use rend3::types::TextureFormat as r3F;
+        Some(match format {
+            k2F::R4G4_UNORM_PACK8
+            | k2F::R4G4B4A4_UNORM_PACK16
+            | k2F::B4G4R4A4_UNORM_PACK16
+            | k2F::R5G6B5_UNORM_PACK16
+            | k2F::B5G6R5_UNORM_PACK16
+            | k2F::R5G5B5A1_UNORM_PACK16
+            | k2F::B5G5R5A1_UNORM_PACK16
+            | k2F::A1R5G5B5_UNORM_PACK16 => return None,
+            k2F::R8_UNORM | k2F::R8_SRGB => {
+                if srgb {
+                    return None;
+                } else {
+                    r3F::R8Unorm
+                }
+            }
+            k2F::R8_SNORM => r3F::R8Snorm,
+            k2F::R8_UINT => r3F::R8Uint,
+            k2F::R8_SINT => r3F::R8Sint,
+            k2F::R8G8_UNORM | k2F::R8G8_SRGB => {
+                if srgb {
+                    return None;
+                } else {
+                    r3F::Rg8Unorm
+                }
+            }
+            k2F::R8G8_SNORM => r3F::Rg8Snorm,
+            k2F::R8G8_UINT => r3F::Rg8Uint,
+            k2F::R8G8_SINT => r3F::Rg8Sint,
+            k2F::R8G8B8_UNORM
+            | k2F::R8G8B8_SNORM
+            | k2F::R8G8B8_UINT
+            | k2F::R8G8B8_SINT
+            | k2F::R8G8B8_SRGB
+            | k2F::B8G8R8_UNORM
+            | k2F::B8G8R8_SNORM
+            | k2F::B8G8R8_UINT
+            | k2F::B8G8R8_SINT
+            | k2F::B8G8R8_SRGB => return None,
+            k2F::R8G8B8A8_UNORM | k2F::R8G8B8A8_SRGB => {
+                if srgb {
+                    r3F::Rgba8UnormSrgb
+                } else {
+                    r3F::Rgba8Unorm
+                }
+            }
+            k2F::R8G8B8A8_SNORM => r3F::Rgba8Snorm,
+            k2F::R8G8B8A8_UINT => r3F::Rgba8Uint,
+            k2F::R8G8B8A8_SINT => r3F::Rgba8Sint,
+            k2F::B8G8R8A8_UNORM | k2F::B8G8R8A8_SRGB => {
+                if srgb {
+                    r3F::Bgra8UnormSrgb
+                } else {
+                    r3F::Bgra8Unorm
+                }
+            }
+            k2F::B8G8R8A8_SNORM | k2F::B8G8R8A8_UINT | k2F::B8G8R8A8_SINT => return None,
+            k2F::A2R10G10B10_UNORM_PACK32
+            | k2F::A2R10G10B10_SNORM_PACK32
+            | k2F::A2R10G10B10_UINT_PACK32
+            | k2F::A2R10G10B10_SINT_PACK32
+            | k2F::A2B10G10R10_UNORM_PACK32
+            | k2F::A2B10G10R10_SNORM_PACK32
+            | k2F::A2B10G10R10_UINT_PACK32
+            | k2F::A2B10G10R10_SINT_PACK32 => return None,
+            k2F::R16_UNORM | k2F::R16_SNORM => return None,
+            k2F::R16_UINT => r3F::R16Uint,
+            k2F::R16_SINT => r3F::R16Sint,
+            k2F::R16_SFLOAT => r3F::R16Float,
+            k2F::R16G16_UNORM | k2F::R16G16_SNORM => return None,
+            k2F::R16G16_UINT => r3F::Rg16Uint,
+            k2F::R16G16_SINT => r3F::Rg16Sint,
+            k2F::R16G16_SFLOAT => r3F::Rg16Float,
+            k2F::R16G16B16_UNORM
+            | k2F::R16G16B16_SNORM
+            | k2F::R16G16B16_UINT
+            | k2F::R16G16B16_SINT
+            | k2F::R16G16B16_SFLOAT => return None,
+            k2F::R16G16B16A16_UNORM | k2F::R16G16B16A16_SNORM => return None,
+            k2F::R16G16B16A16_UINT => r3F::Rgba16Uint,
+            k2F::R16G16B16A16_SINT => r3F::Rgba16Sint,
+            k2F::R16G16B16A16_SFLOAT => r3F::Rgba16Float,
+            k2F::R32_UINT => r3F::R32Uint,
+            k2F::R32_SINT => r3F::R32Sint,
+            k2F::R32_SFLOAT => r3F::R32Float,
+            k2F::R32G32_UINT => r3F::Rg32Uint,
+            k2F::R32G32_SINT => r3F::Rg32Sint,
+            k2F::R32G32_SFLOAT => r3F::Rg32Float,
+            k2F::R32G32B32_UINT | k2F::R32G32B32_SINT | k2F::R32G32B32_SFLOAT => return None,
+            k2F::R32G32B32A32_UINT => r3F::Rgba32Uint,
+            k2F::R32G32B32A32_SINT => r3F::Rgba32Sint,
+            k2F::R32G32B32A32_SFLOAT => r3F::Rgba32Float,
+            k2F::R64_UINT
+            | k2F::R64_SINT
+            | k2F::R64_SFLOAT
+            | k2F::R64G64_UINT
+            | k2F::R64G64_SINT
+            | k2F::R64G64_SFLOAT
+            | k2F::R64G64B64_UINT
+            | k2F::R64G64B64_SINT
+            | k2F::R64G64B64_SFLOAT
+            | k2F::R64G64B64A64_UINT
+            | k2F::R64G64B64A64_SINT
+            | k2F::R64G64B64A64_SFLOAT => return None,
+            k2F::B10G11R11_UFLOAT_PACK32 => r3F::Rg11b10Float,
+            k2F::E5B9G9R9_UFLOAT_PACK32 => r3F::Rgb9e5Ufloat,
+            k2F::D16_UNORM => return None,
+            k2F::X8_D24_UNORM_PACK32 => r3F::Depth24Plus,
+            k2F::D32_SFLOAT => r3F::Depth32Float,
+            k2F::S8_UINT | k2F::D16_UNORM_S8_UINT => return None,
+            k2F::D24_UNORM_S8_UINT => r3F::Depth24PlusStencil8,
+            k2F::D32_SFLOAT_S8_UINT => return None,
+            k2F::BC1_RGB_UNORM_BLOCK
+            | k2F::BC1_RGB_SRGB_BLOCK
+            | k2F::BC1_RGBA_UNORM_BLOCK
+            | k2F::BC1_RGBA_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Bc1RgbaUnormSrgb
+                } else {
+                    r3F::Bc1RgbaUnorm
+                }
+            }
+            k2F::BC2_UNORM_BLOCK | k2F::BC2_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Bc2RgbaUnormSrgb
+                } else {
+                    r3F::Bc2RgbaUnorm
+                }
+            }
+            k2F::BC3_UNORM_BLOCK | k2F::BC3_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Bc3RgbaUnormSrgb
+                } else {
+                    r3F::Bc3RgbaUnorm
+                }
+            }
+            k2F::BC4_UNORM_BLOCK => r3F::Bc4RUnorm,
+            k2F::BC4_SNORM_BLOCK => r3F::Bc4RSnorm,
+            k2F::BC5_UNORM_BLOCK => r3F::Bc5RgUnorm,
+            k2F::BC5_SNORM_BLOCK => r3F::Bc5RgSnorm,
+            k2F::BC6H_UFLOAT_BLOCK => r3F::Bc6hRgbUfloat,
+            k2F::BC6H_SFLOAT_BLOCK => r3F::Bc6hRgbSfloat,
+            k2F::BC7_UNORM_BLOCK | k2F::BC7_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Bc7RgbaUnormSrgb
+                } else {
+                    r3F::Bc7RgbaUnorm
+                }
+            }
+            k2F::ETC2_R8G8B8_UNORM_BLOCK | k2F::ETC2_R8G8B8_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Etc2RgbUnormSrgb
+                } else {
+                    r3F::Etc2RgbUnorm
+                }
+            }
+            k2F::ETC2_R8G8B8A1_UNORM_BLOCK | k2F::ETC2_R8G8B8A1_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Etc2RgbA1UnormSrgb
+                } else {
+                    r3F::Etc2RgbA1Unorm
+                }
+            }
+            k2F::ETC2_R8G8B8A8_UNORM_BLOCK | k2F::ETC2_R8G8B8A8_SRGB_BLOCK => return None,
+            k2F::EAC_R11_UNORM_BLOCK => r3F::EacRUnorm,
+            k2F::EAC_R11_SNORM_BLOCK => r3F::EacRSnorm,
+            k2F::EAC_R11G11_UNORM_BLOCK => r3F::EacRgUnorm,
+            k2F::EAC_R11G11_SNORM_BLOCK => r3F::EacRgSnorm,
+            k2F::ASTC_4x4_UNORM_BLOCK | k2F::ASTC_4x4_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc4x4RgbaUnormSrgb
+                } else {
+                    r3F::Astc4x4RgbaUnorm
+                }
+            }
+            k2F::ASTC_5x4_UNORM_BLOCK | k2F::ASTC_5x4_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc5x4RgbaUnormSrgb
+                } else {
+                    r3F::Astc5x4RgbaUnorm
+                }
+            }
+            k2F::ASTC_5x5_UNORM_BLOCK | k2F::ASTC_5x5_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc5x5RgbaUnormSrgb
+                } else {
+                    r3F::Astc5x5RgbaUnorm
+                }
+            }
+            k2F::ASTC_6x5_UNORM_BLOCK | k2F::ASTC_6x5_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc6x5RgbaUnormSrgb
+                } else {
+                    r3F::Astc6x5RgbaUnorm
+                }
+            }
+            k2F::ASTC_6x6_UNORM_BLOCK | k2F::ASTC_6x6_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc6x6RgbaUnormSrgb
+                } else {
+                    r3F::Astc6x6RgbaUnorm
+                }
+            }
+            k2F::ASTC_8x5_UNORM_BLOCK | k2F::ASTC_8x5_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc8x5RgbaUnormSrgb
+                } else {
+                    r3F::Astc8x5RgbaUnorm
+                }
+            }
+            k2F::ASTC_8x6_UNORM_BLOCK | k2F::ASTC_8x6_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc8x6RgbaUnormSrgb
+                } else {
+                    r3F::Astc8x6RgbaUnorm
+                }
+            }
+            k2F::ASTC_8x8_UNORM_BLOCK | k2F::ASTC_8x8_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc8x8RgbaUnormSrgb
+                } else {
+                    r3F::Astc8x8RgbaUnorm
+                }
+            }
+            k2F::ASTC_10x5_UNORM_BLOCK | k2F::ASTC_10x5_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc10x5RgbaUnormSrgb
+                } else {
+                    r3F::Astc10x5RgbaUnorm
+                }
+            }
+            k2F::ASTC_10x6_UNORM_BLOCK | k2F::ASTC_10x6_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc10x6RgbaUnormSrgb
+                } else {
+                    r3F::Astc10x6RgbaUnorm
+                }
+            }
+            k2F::ASTC_10x8_UNORM_BLOCK | k2F::ASTC_10x8_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc10x8RgbaUnormSrgb
+                } else {
+                    r3F::Astc10x8RgbaUnorm
+                }
+            }
+            k2F::ASTC_10x10_UNORM_BLOCK | k2F::ASTC_10x10_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc10x10RgbaUnormSrgb
+                } else {
+                    r3F::Astc10x10RgbaUnorm
+                }
+            }
+            k2F::ASTC_12x10_UNORM_BLOCK | k2F::ASTC_12x10_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc12x10RgbaUnormSrgb
+                } else {
+                    r3F::Astc12x10RgbaUnorm
+                }
+            }
+            k2F::ASTC_12x12_UNORM_BLOCK | k2F::ASTC_12x12_SRGB_BLOCK => {
+                if srgb {
+                    r3F::Astc12x12RgbaUnormSrgb
+                } else {
+                    r3F::Astc12x12RgbaUnorm
+                }
+            }
+            _ => return None,
+        })
     }
 }