+++ rend3-pbr/src/culling/gpu.rs
@@ -11,7 +11,7 @@ use wgpu::{
     util::{BufferInitDescriptor, DeviceExt},
     BindGroupDescriptor, BindGroupEntry, BindGroupLayout, BindGroupLayoutDescriptor, BindGroupLayoutEntry, BindingType,
     Buffer, BufferBindingType, BufferDescriptor, BufferUsages, CommandEncoder, ComputePassDescriptor, ComputePipeline,
-    ComputePipelineDescriptor, Device, PipelineLayoutDescriptor, PushConstantRange, RenderPass,
+    ComputePipelineDescriptor, Device, PipelineLayoutDescriptor, PushConstantRange, RenderPass, ShaderModuleDescriptor,
     ShaderModuleDescriptorSpirV, ShaderStages,
 };
 
@@ -74,6 +74,8 @@ pub struct GpuCuller {
 }
 impl GpuCuller {
     pub fn new(device: &Device) -> Self {
+        profiling::scope!("GpuCuller::new");
+
         let atomic_bgl = device.create_bind_group_layout(&BindGroupLayoutDescriptor {
             label: Some("atomic culling pll"),
             entries: &[
@@ -216,35 +218,25 @@ impl GpuCuller {
             })
         };
 
-        let prefix_cull_sm = unsafe {
-            device.create_shader_module_spirv(&ShaderModuleDescriptorSpirV {
-                label: Some("cull-prefix-cull"),
-                source: wgpu::util::make_spirv_raw(
-                    SPIRV_SHADERS.get_file("cull-prefix-cull.comp.spv").unwrap().contents(),
-                ),
-            })
-        };
+        let prefix_cull_sm = device.create_shader_module(&ShaderModuleDescriptor {
+            label: Some("cull-prefix-cull"),
+            source: wgpu::util::make_spirv(SPIRV_SHADERS.get_file("cull-prefix-cull.comp.spv").unwrap().contents()),
+        });
 
-        let prefix_sum_sm = unsafe {
-            device.create_shader_module_spirv(&ShaderModuleDescriptorSpirV {
-                label: Some("cull-prefix-sum"),
-                source: wgpu::util::make_spirv_raw(
-                    SPIRV_SHADERS.get_file("cull-prefix-sum.comp.spv").unwrap().contents(),
-                ),
-            })
-        };
+        let prefix_sum_sm = device.create_shader_module(&ShaderModuleDescriptor {
+            label: Some("cull-prefix-sum"),
+            source: wgpu::util::make_spirv(SPIRV_SHADERS.get_file("cull-prefix-sum.comp.spv").unwrap().contents()),
+        });
 
-        let prefix_output_sm = unsafe {
-            device.create_shader_module_spirv(&ShaderModuleDescriptorSpirV {
-                label: Some("cull-prefix-output"),
-                source: wgpu::util::make_spirv_raw(
-                    SPIRV_SHADERS
-                        .get_file("cull-prefix-output.comp.spv")
-                        .unwrap()
-                        .contents(),
-                ),
-            })
-        };
+        let prefix_output_sm = device.create_shader_module(&ShaderModuleDescriptor {
+            label: Some("cull-prefix-output"),
+            source: wgpu::util::make_spirv(
+                SPIRV_SHADERS
+                    .get_file("cull-prefix-output.comp.spv")
+                    .unwrap()
+                    .contents(),
+            ),
+        });
 
         let atomic_pipeline = device.create_compute_pipeline(&ComputePipelineDescriptor {
             label: Some("atomic culling pl"),
@@ -291,7 +283,7 @@ impl GpuCuller {
         if let Some(sorting) = args.sort {
             profiling::scope!("Sorting");
 
-            let camera_location = args.camera.get_data().location;
+            let camera_location = args.camera.location().into();
 
             match sorting {
                 Sorting::FrontToBack => {