+++ rend3/src/resources/camera.rs
@@ -1,12 +1,12 @@
 use crate::types::{Camera, CameraProjection};
-use glam::{Mat3A, Mat4, Vec3, Vec3A};
+use glam::{Mat4, Vec3};
 
 /// Manages the camera's location and projection settings.
 #[derive(Debug, Clone)]
 pub struct CameraManager {
     orig_view: Mat4,
-    view: Mat4,
     proj: Mat4,
+    inv_view: Mat4,
     data: Camera,
     aspect_ratio: f32,
 }
@@ -14,15 +14,16 @@ impl CameraManager {
     /// Builds a new camera, using the given aspect ratio. If no aspect ratio is given
     /// it is assumed that no aspect ratio scaling should be done.
     pub fn new(data: Camera, aspect_ratio: Option<f32>) -> Self {
+        profiling::scope!("CameraManager::new");
+
         let aspect_ratio = aspect_ratio.unwrap_or(1.0);
         let proj = compute_projection_matrix(data, aspect_ratio);
-        let view = compute_view_matrix(data);
         let orig_view = compute_origin_matrix(data);
 
         Self {
             orig_view,
-            view,
             proj,
+            inv_view: data.view.inverse(),
             data,
             aspect_ratio,
         }
@@ -40,8 +41,8 @@ impl CameraManager {
 
     pub fn set_aspect_data(&mut self, data: Camera, aspect_ratio: f32) {
         self.proj = compute_projection_matrix(data, self.aspect_ratio);
-        self.view = compute_view_matrix(data);
         self.orig_view = compute_origin_matrix(data);
+        self.inv_view = data.view.inverse();
         self.data = data;
         self.aspect_ratio = aspect_ratio;
     }
@@ -51,11 +52,11 @@ impl CameraManager {
     }
 
     pub fn view(&self) -> Mat4 {
-        self.view
+        self.data.view
     }
 
     pub fn view_proj(&self) -> Mat4 {
-        self.proj * self.view
+        self.proj * self.data.view
     }
 
     pub fn origin_view_proj(&self) -> Mat4 {
@@ -65,26 +66,10 @@ impl CameraManager {
     pub fn proj(&self) -> Mat4 {
         self.proj
     }
-}
 
-fn compute_look_offset(data: Camera) -> Vec3A {
-    match data.projection {
-        CameraProjection::Projection { pitch, yaw, .. } => {
-            let starting = Vec3A::Z;
-            Mat3A::from_euler(glam::EulerRot::YXZ, yaw, pitch, 0.0) * starting
-        }
-        CameraProjection::Orthographic { direction, .. } => direction,
+    pub fn location(&self) -> Vec3 {
+        self.inv_view.w_axis.truncate()
     }
-}
-
-fn compute_view_matrix(data: Camera) -> Mat4 {
-    let look_offset = compute_look_offset(data);
-
-    Mat4::look_at_lh(
-        Vec3::from(data.location),
-        Vec3::from(data.location + look_offset),
-        Vec3::Y,
-    )
 }
 
 fn compute_projection_matrix(data: Camera, aspect_ratio: f32) -> Mat4 {
@@ -99,9 +84,9 @@ fn compute_projection_matrix(data: Camera, aspect_ratio: f32) -> Mat4 {
     }
 }
 
-// This is horribly inefficient but is called like once a frame.
 fn compute_origin_matrix(data: Camera) -> Mat4 {
-    let look_offset = compute_look_offset(data);
+    let mut view = data.view;
 
-    Mat4::look_at_lh(Vec3::ZERO, Vec3::from(look_offset), Vec3::Y)
+    view.w_axis = glam::Vec4::W;
+    view
 }